+++ src/app/cli/src/commands/pull_command.rs
@@ -110,7 +110,6 @@ impl PullCommand {
             self.search_svc.clone(),
             self.refs.clone(),
         )
-        .map_ok(|dataset_ref_any| dataset_ref_any)
         .try_collect()
         .await?;