+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Fixed `mapboxgl` dependency issue in Jupyter image
+
 ## [0.163.0] - 2024-03-01
 ### Changed
 - Simplified flow statuses within Flow System (no more Queued or Scheduled status)