+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.164.0] - 2024-03-06
 ### Added
 - Added support of wildcard patterns for `kamu pull` and `kamu push` commands
 - Added `{dataset}/tail` and `/query` REST API endpoints